refactor draenor script loading
