+++ src/server/scripts/Draenor/zone_frostfire_ridge.cpp
