+++ src/server/scripts/Draenor/zone_spires_of_arak.cpp
@@ -1,0 +1,282 @@
+#include "ScriptMgr.h"
+#include "ScriptedCreature.h"
+
+enum eSpells
+{
+    // ruhamar
+    SPELL_SOLAR_RADIATION       = 167708,
+    SPELL_SHARP_BEAK            = 167614, // 12
+    SPELL_SOLAR_BREATH          = 167679, // 29
+    SPELL_LOOSE_QUILLS          = 167647, // 115 (fly)
+    SPELL_BLOOD_FEATHER         = 167625, // 15 + script adds
+    SPELL_SLEEP_ADDS            = 160641,
+    SPELL_FIX_TARGET            = 167757,
+
+    EVENT_GROUP_GROUND          = 1
+};
+
+// 83746
+class boss_ruhamar : public CreatureScript
+{
+public:
+    boss_ruhamar() : CreatureScript("boss_ruhamar") { }
+
+    struct boss_ruhamarAI : public ScriptedAI
+    {
+        boss_ruhamarAI(Creature* creature) : ScriptedAI(creature), summons(me) {}
+
+        EventMap events;
+        SummonList summons;
+        uint32 healthPct;
+        Position oldpos;
+
+        void Reset()
+        {
+            me->GetMotionMaster()->MovePath(439156, true);
+            me->SetCanFly(true);
+            me->SetDisableGravity(true);
+            me->SetByteFlag(UNIT_FIELD_BYTES_1, 3, UNIT_BYTE1_FLAG_ALWAYS_STAND | UNIT_BYTE1_FLAG_HOVER);
+            me->SetReactState(REACT_AGGRESSIVE);
+
+            events.Reset();
+            summons.DespawnAll();
+            healthPct = 98;
+        }
+
+        void EnterCombat(Unit* unit)
+        {
+            me->SetDisableGravity(false);
+            me->RemoveByteFlag(UNIT_FIELD_BYTES_1, 3,  UNIT_BYTE1_FLAG_ALWAYS_STAND | UNIT_BYTE1_FLAG_HOVER);
+            me->CastSpell(me, SPELL_SOLAR_RADIATION);
+            events.RescheduleEvent(EVENT_1, 12000, EVENT_GROUP_GROUND);
+            events.RescheduleEvent(EVENT_2, 29000, EVENT_GROUP_GROUND);
+            events.RescheduleEvent(EVENT_3, 85000);
+        }
+
+        void JustSummoned(Creature* summon)
+        {
+            summons.Summon(summon);
+            DoZoneInCombat(summon, 150.0f);
+        }
+
+        void DamageTaken(Unit* /*attacker*/, uint32& damage, DamageEffectType dmgType)
+        {
+            if (me->HealthBelowPct(healthPct))
+            {
+                healthPct -= urand(2,4);
+                events.RescheduleEvent(EVENT_4, 10, EVENT_GROUP_GROUND);
+            }
+        }
+
+        void OnApplyOrRemoveAura(uint32 spellId, AuraRemoveMode mode, bool apply)
+        {
+            if (!apply && spellId == SPELL_LOOSE_QUILLS)
+                me->GetMotionMaster()->MovePoint(2, oldpos.GetPositionX(), oldpos.GetPositionY(), oldpos.GetPositionZ());
+        }
+
+        void MovementInform(uint32 type, uint32 point)
+        {
+            if (type == WAYPOINT_MOTION_TYPE)
+                me->SetHomePosition(me->GetPositionX(), me->GetPositionY(), me->GetPositionZ(), me->GetOrientation());
+
+            if (type != POINT_MOTION_TYPE)
+                return;
+
+            me->GetMotionMaster()->MovementExpired();
+            if (point == 1)
+                me->CastSpell(me, SPELL_LOOSE_QUILLS);
+            else if (point == 2)
+            {
+                me->SetCanFly(false);
+                me->SetDisableGravity(false);
+                me->RemoveByteFlag(UNIT_FIELD_BYTES_1, 3, UNIT_BYTE1_FLAG_ALWAYS_STAND | UNIT_BYTE1_FLAG_HOVER);
+                me->SetReactState(REACT_AGGRESSIVE);
+
+                events.RescheduleEvent(EVENT_1, 12000, EVENT_GROUP_GROUND);
+                events.RescheduleEvent(EVENT_2, 29000, EVENT_GROUP_GROUND);
+                events.RescheduleEvent(EVENT_3, 85000);
+            }
+        }
+
+
+        void UpdateAI(uint32 diff)
+        {
+            if (!UpdateVictim() || me->HasUnitState(UNIT_STATE_CASTING))
+                return;
+
+            events.Update(diff);
+
+            if (uint32 eventId = events.ExecuteEvent())
+            {
+				switch (eventId)
+				{
+				case EVENT_1:
+					DoCastVictim(SPELL_SHARP_BEAK);
+					events.RescheduleEvent(EVENT_1, 12000, EVENT_GROUP_GROUND);
+					break;
+				case EVENT_2:
+					DoCastVictim(SPELL_SOLAR_BREATH);
+					events.RescheduleEvent(EVENT_2, 29000, EVENT_GROUP_GROUND);
+					break;
+				case EVENT_3:
+				{
+					events.CancelEventGroup(EVENT_GROUP_GROUND);
+
+					me->SetCanFly(true);
+					me->SetDisableGravity(true);
+					me->SetByteFlag(UNIT_FIELD_BYTES_1, 3, UNIT_BYTE1_FLAG_ALWAYS_STAND | UNIT_BYTE1_FLAG_HOVER);
+					me->SetReactState(REACT_PASSIVE);
+					me->AttackStop();
+					Position pos;
+					pos.Relocate(me);
+					oldpos = pos;
+					pos.m_positionZ += 10.0f;
+					me->GetMotionMaster()->MovePoint(1, pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ());
+					break;
+				}
+                    case EVENT_4:
+                        DoCastVictim(SPELL_BLOOD_FEATHER);
+                        break;
+                }
+            }
+
+            DoMeleeAttackIfReady();
+        }
+    };
+
+    CreatureAI* GetAI(Creature* creature) const
+    {
+        return new boss_ruhamarAI(creature);
+    }
+};
+
+// 83769
+class npc_energized_phoenix : public CreatureScript
+{
+public:
+    npc_energized_phoenix() : CreatureScript("npc_energized_phoenix") { }
+
+    struct npc_energized_phoenixAI : public ScriptedAI
+    {
+        npc_energized_phoenixAI(Creature* creature) : ScriptedAI(creature) {}
+
+        EventMap events;
+        bool moveComplete;
+        uint32 moveTimer;
+		ObjectGuid targetMove;
+
+        void Reset()
+        {
+            events.Reset();
+        }
+
+        void EnterCombat(Unit* unit)
+        {
+        }
+
+
+        void DamageTaken(Unit* /*attacker*/, uint32& damage, DamageEffectType dmgType)
+        {
+            if (damage >= me->GetHealth())
+            {
+                damage = 0;
+
+                me->CastStop();
+                me->SetHealth(me->GetMaxHealth());
+                me->CastSpell(me, 167629, true);
+                me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE | UNIT_FLAG_IMMUNE_TO_PC | UNIT_FLAG_IMMUNE_TO_NPC | UNIT_FLAG_REMOVE_CLIENT_CONTROL);
+                me->CastSpell(me, SPELL_SLEEP_ADDS, true);
+                me->GetMotionMaster()->MovementExpired(true);
+                me->SetDisplayId(66325);
+                me->AttackStop();
+                me->SetReactState(REACT_PASSIVE);
+            }
+        }
+
+        void IsSummonedBy(Unit* owner)
+        {
+            if (owner->GetGUID() != me->GetGUID())
+            {
+                std::list<Creature*> adds;
+                GetCreatureListWithEntryInGrid(adds, me, me->GetEntry(), 5.0f);
+                if (!adds.empty())
+                    for (std::list<Creature*>::iterator itr = adds.begin(); itr != adds.end(); ++itr)
+                    {
+                        if ((*itr)->HasAura(SPELL_SLEEP_ADDS))
+                        {
+                            (*itr)->RemoveAurasDueToSpell(SPELL_SLEEP_ADDS);
+                            (*itr)->AI()->IsSummonedBy((*itr));
+                        }
+                    }
+            }
+
+            me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE | UNIT_FLAG_IMMUNE_TO_PC | UNIT_FLAG_IMMUNE_TO_NPC | UNIT_FLAG_REMOVE_CLIENT_CONTROL);
+            me->SetDisplayId(me->GetNativeDisplayId());
+
+            moveComplete = false;
+            moveTimer = 2000;
+
+            std::list<Player*> targets;
+            targets.clear();
+            GetPlayerListInGrid(targets, me, 70.0f);
+            if (!targets.empty())
+            {
+                std::list<Player*>::const_iterator itr = targets.begin();
+                std::advance(itr, urand(0, targets.size() - 1));
+                if (*itr)
+                    targetMove = (*itr)->GetGUID();
+
+                me->GetMotionMaster()->MoveFollow((*itr), PET_FOLLOW_DIST, (*itr)->GetFollowAngle());
+                me->CastSpell((*itr), SPELL_FIX_TARGET);
+            }
+            else
+                me->DespawnOrUnsummon();
+        }
+
+
+        void UpdateAI(uint32 diff)
+        {
+            if (!moveComplete)
+            {
+                if (moveTimer <= diff)
+                {
+                    Player* target = ObjectAccessor::GetPlayer(*me, targetMove);
+                    if (!target)
+                    {
+                        IsSummonedBy(me); // tru refind
+                        return;
+                    }
+
+                    if (me->GetDistance(target) < 3.0f)
+                    {
+                        moveComplete = true;
+
+                        me->CastStop();
+                        me->SetHealth(me->GetMaxHealth());
+                        me->CastSpell(me, 167629, true);
+                        me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE | UNIT_FLAG_IMMUNE_TO_PC | UNIT_FLAG_IMMUNE_TO_NPC | UNIT_FLAG_REMOVE_CLIENT_CONTROL);
+                        me->CastSpell(me, SPELL_SLEEP_ADDS, true);
+                        me->GetMotionMaster()->MovementExpired(true);
+                        me->SetDisplayId(66325);
+                        me->AttackStop();
+                        me->SetReactState(REACT_PASSIVE);
+                    }
+                    moveTimer = 1000;
+                }
+                else
+                    moveTimer -= diff;
+            }
+        }
+    };
+
+    CreatureAI* GetAI(Creature* creature) const
+    {
+        return new npc_energized_phoenixAI(creature);
+    }
+};
+
+void AddSC_wod_spires_of_arak()
+{
+    new boss_ruhamar();
+    new npc_energized_phoenix();
+}
